requirements.txt
streamlit
tensorflow==2.13.0
numpy
pandas
pillow
openpyxl
